+++ src/domain/entities/UserShortUrl.ts
@@ -1,0 +1,8 @@
+export interface IUserShortUrl {
+  userId: string;
+  shortUrlId: string;
+  click_count: number;
+  createdAt: Date;
+  updatedAt: Date;
+  deletedAt?: Date;
+}