+++ src/infrastructure/http/routes/index.ts
@@ -1,8 +1,10 @@
 import { Router } from "express";
 import { shortUrlsRoutes } from "./short-urls.routes";
+import { redirectRoutes } from "./redirect.routes";
 
 const router = Router();
 
 router.use("/short-urls", shortUrlsRoutes);
+router.use("/", redirectRoutes);
 
 export { router };