+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [unreleased]
+
+### Added
+
+- Redirect route to redirect the shortened URL to the original URL.
+- Error code for better identification of errors.
+
+### Changed
+
+- Table name from `short_url` to `short_urls`.
+
 ## [0.4.0] - 2025-04-10
 
 ### Added
@@ -14,7 +25,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
-- Change name table from `short_urls` to `short_url`.
+- Table name from `short_urls` to `short_url`.
 
 ## [v0.3.0] - 2025-04-10